added frontend

+++ backend/controllers/goalController.js
@@ -1,11 +1,13 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
+
 
 // @desc      Get goals
 // @route     GET /api/goals
 // @access    Private
 const getGoals = asyncHandler(async (req, res) => {
-       const goals = await Goal.find();
+       const goals = await Goal.find({user: req.user.id});
        res.status(200).json(goals);
 })
 
@@ -22,7 +24,8 @@ const createGoal = asyncHandler(async (req, res) => {
        }
 
        const goal = await Goal.create({
-              text: req.body.text
+              text: req.body.text,
+              user: req.user.id
        })
        res.status(200).json({status: true, goal: goal});
 })
@@ -33,10 +36,23 @@ const createGoal = asyncHandler(async (req, res) => {
 // @access    Private
 const updateGoal = asyncHandler(async (req, res) => {
        const goal = await Goal.findById(req.params.id);
+       const user = await User.findById(req.user.id);
+
 
        if(!goal){
               res.status(400);
               throw new Error('Goal not found');
+       }
+       // check for user
+       if(!user){
+              res.status(401);
+              throw new Error('User not found')
+       }
+
+       // Make sure the logged in user matches the goal user
+       if(goal.user.toString() !== req.user.id){
+              res.status(401);
+              throw new Error('User not author');
        }
        const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
               new: true
@@ -50,9 +66,23 @@ const updateGoal = asyncHandler(async (req, res) => {
 // @access    Private
 const deleteGoal = asyncHandler(async (req, res) => {
        const goal = await Goal.findById(req.params.id);
+       const user = await User.findById(req.user.id);
+
+
        if(!goal){
               res.status(400);
               throw new Error('Goal not found');
+       }
+       // check for user
+       if(!user){
+              res.status(401);
+              throw new Error('User not found')
+       }
+
+       // Make sure the logged in user matches the goal user
+       if(goal.user.toString() !== req.user.id){
+              res.status(401);
+              throw new Error('User not author');
        }
        await Goal.findByIdAndDelete(req.params.id);
        // or
